Server runs with more modern design
