+++ package.json
@@ -8,8 +8,8 @@
     "clean": "rm -rf dist/*",
     "dev": "NODE_NO_WARNINGS=1 yarn run server:dev",
     "build": "vite build",
-    "server": "node ./server/index.js",
-    "server:dev": "tsx ./server/index.ts",
+    "server": "tsx ./server/index.ts",
+    "server:dev": "tsx --env-file=.env ./server/index.ts",
     "format": "prettier --write src packages",
     "server:prod": "NODE_NO_WARNINGS=1 NODE_ENV=production yarn run server",
     "dev:storybook": "yarn workspace @macrostrat/storybook run dev",
